Clippy

+++ src/time.rs
@@ -94,7 +94,7 @@ fn debug_tick_system(mut tick_counter: ResMut<TickCounter>) {
         let tick = tick_counter.tick;
 
         let now = Instant::now();
-        let avg_dur = now.duration_since(tick_counter.since) / TickCounter::DIAGNOSTIC_ITERS as u32;
+        let avg_dur = now.duration_since(tick_counter.since) / TickCounter::DIAGNOSTIC_ITERS;
         tick_counter.since = now;
 
         info!(tick, ?avg_dur, "Tick");
